requirements.txt
fastapi
uvicorn

#for database   #new
sqlalchemy         
psycopg2

#for loading environment variables  #new
python-dotenv

#for your ML
pandas
numpy